+++ src/os2datascanner/projects/admin/core/management/commands/run_background_jobs.py
@@ -24,7 +24,7 @@ REGISTRY = CollectorRegistry()
 JOB_STATE = Enum('background_job_status',
                  'Shows the current/last state of given background job',
                  states=['waiting', 'running', 'cancelling',
-                         'finished', 'cancelled', 'failed'],
+                         'finished', 'cancelled', 'failed', 'finished with warnings'],
                  labelnames=['JobLabel', 'OrgSlug', 'OrgId'])
 REGISTRY.register(JOB_STATE)  # Register the metric
 PUSHGATEWAY_HOST = settings.PUSHGATEWAY_HOST
@@ -178,7 +178,8 @@ class Command(BaseCommand):
                                 job.save()
                             elif job.exec_state not in (
                                     JobState.FAILED,
-                                    JobState.CANCELLED):
+                                    JobState.CANCELLED,
+                                    JobState.FINISHED_WITH_WARNINGS):
                                 job.exec_state = JobState.FINISHED
                                 job.save()
                             publish_job_state(job)