+++ src/os2datascanner/projects/admin/import_services/utils.py
@@ -47,7 +47,8 @@ def start_ldap_import(ldap_conf: LDAPConfig):
     if not latest_importjob \
             or latest_importjob.exec_state == JobState.FINISHED \
             or latest_importjob.exec_state == JobState.FAILED \
-            or latest_importjob.exec_state == JobState.CANCELLED:
+            or latest_importjob.exec_state == JobState.CANCELLED \
+            or latest_importjob.exec_state == JobState.FINISHED_WITH_WARNINGS:
         LDAPImportJob.objects.create(
             realm=realm
         )